Scraper/requirements.txt
Flask==1.1.2
flask_apscheduler
newspaper3k
dnspython
pymongo
dacite
gunicorn==20.0.4